refactor: Direction을 상위 패키지로 이동
